Remove copying metadata on disk from output repos
